Update README and refactor error diffusion logic; add new dithering commands
